organized class additions
