+++ android_files/app/src/main/java/com/example/qradventure/ui/leaderboard/PlayerListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.qradventure;
+package com.example.qradventure.ui.leaderboard;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,8 +7,9 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.qradventure.R;
 
 import java.util.ArrayList;
